remote_access_tool/requirements.txt
psycopg2-binary==2.9.6
paramiko==2.11.0
python-dotenv==1.0.0